wzry-main/wzry_offline/wzry_sampler/requirements.txt
numpy~=2.2.0
opencv-python~=4.10.0.84
psutil~=6.1.0
torch~=2.5.1
gym~=0.26.2
pynput~=1.7.7
openai~=1.57.4
PyQt5~=5.15.11
pillow~=11.0.0
ultralytics~=8.3.50
pytesseract~=0.3.13
pywin32==308